requirements.txt
discord.py==2.4.0
g4f==0.3.2.4
python-dotenv==1.0.1
curl_cffi==0.7.1
nest_asyncio==1.6.0